# Minhash improvements