The main change here are consolidating two lines into one by removing tokens.

Besides simplification, this helps with memory consumption since `set` consumes much more memory (due to the ratio between the content and hashes, almost twice a list)
Since we are using the `set` datastructure for deduplication and is later converted into a list, we do not have to preserve that structure.

Although I have witnessed some speed acceleration as well, it is not consistent and seems to be within statistical variance. I did not witness any regressions though.