@ChenghaoMou 2023-08-13
I don't understand the justification here.

A document usually has **N** shingles and **M** unique shingles (N >> M). In the PR, you create an np.array with hash_func called for each one of the **N** shingles first. I can't see how this changes anything in any meaningful way. 

One could, for example:
```
hashvalues: np.ndarray = np.array([
    hash_func(token) for token in {
        bytes(" ".join(t), "utf-8") 
        for t in ngrams(NON_ALPHA.split(content.lower()), 
        ngram_size, min_length)
    }
], dtype=dtype)
```
skip the variable by creating the set on the fly if memory is concerned.