@ChenghaoMou 2023-08-13
I tested the PR's embed_func with the file content from `text_dedup/minhash.py`. Here is the result:

```
Before:
3.95 ms ± 65.1 µs per loop (mean ± std. dev. of 3 runs, 1,000 loops each)

After:
4.18 ms ± 24.4 µs per loop (mean ± std. dev. of 3 runs, 1,000 loops each)

Skip `tokens` but keep set:
3.77 ms ± 191 µs per loop (mean ± std. dev. of 3 runs, 1,000 loops each)
```

Code snippet to reproduce:

```
from pathlib import Path

content = Path("/Users/chenghao/Downloads/text-dedup/text_dedup/minhash.py").read_text()
idx = 0
num_perm = 250
ngram_size = 5
hashranges = [(i, i + 25) for i in range(0, 250, 25)]
max_hash = np.uint32((1 << 32) - 1)
modulo_prime = np.uint32((1 << 32) - 5)
PERMUTATIONS = np.array(
    [
        (
            RNG.randint(1, np.uint32((1 << 32) - 5), dtype=np.uint32),
            RNG.randint(0, np.uint32((1 << 32) - 5), dtype=np.uint32),
        )
        for _ in range(num_perm)
    ],
    dtype=np.uint32,
).T

%%timeit -n 1000 -r 3
res = embed_func(
    content, 
    idx,
    num_perm=num_perm, 
    ngram_size=ngram_size, 
    min_length=0, 
    hashranges=hashranges,
    permutations=PERMUTATIONS, 
    hash_func=xxh3_32hash,
    dtype=np.uint32, 
    max_hash=max_hash, 
    modulo_prime=modulo_prime
)
```