@chris-ha458 2023-08-13
Thanks for the reproduction code. it will come in handy going forward.
I might even rearrange it into some microbenchmark and add it as a future PR.

As for my part of the testing. (PC is AMD Ryzen 7950x)
(for mor extensive testing i've done this)
`%%timeit -n 2000-r 7`

```
Original : 
3.24 ms ± 73 µs per loop (mean ± std. dev. of 7 runs, 2,000 loops each)
This PR : 
3.24 ms ± 41.1 µs per loop (mean ± std. dev. of 7 runs, 2,000 loops each)
This PR but operating on list directly without making np.array in the middle.:
3.23 ms ± 23 µs per loop (mean ± std. dev. of 7 runs, 2,000 loops each)
ChenghaoMa Version Suggested in PR : 
3.24 ms ± 50.2 µs per loop (mean ± std. dev. of 7 runs, 2,000 loops each)
np.fromiter (Not ever making a list at all and building up with iters)
3.29 ms ± 58.4 µs per loop (mean ± std. dev. of 7 runs, 2,000 loops each)
```
What minor differences shown above are statistical variances.
I've tested much more under different settings but each version all come within 3.2ms~3.3ms upon repeated testing

It seems that under these test all these choices are neutral at my machine, it could very well be a regression in others. 
I feel that whatever gains I might have witnessed on operating on OSCAR datasets might have been statistical variances as well.
Multiple other methods I've thought of do not show any tangible reproducible benefits either.
My motivation to exploit vectorized advanced numpy functions was a great underestimation of python native implementations.

One side benefit i was hoping for was enhanced readability but that is also controversial.
I think atleast for now, reverting `embed_func` back to the original version is the best way to go.

An underlying reason I have been trying to use np native functions and datastructures as much as possible was to rewrite as many functions as possible to support vectorization, but it seems to overcomplicate the code and make it difficult to read and understand with questionable benefits at the moment.

Also, even if i achieve some level of vectorization that benefits the above micro bench, it is unclear if it would be beneficial when instance of `embed_func` is also launched in a multiprocessing manner with `dataset.map()`


I added a revert commit to revert the `tokens` and `hashvalues` changes.